# hisabbot/openapi.json
{"openapi":"3.1.0","info":{"title":"HisabBot API","version":"3.0"},"paths":{"/voice/transcribe":{"post":{"tags":["voice"],"summary":"Transcribe","description":"Accept a browser audio blob (webm/wav) and return transcribed Roman Urdu text.\n\nThe browser records via MediaRecorder API and POSTs the blob here.\nGroq Whisper handles Roman Urdu / mixed Urdu-English well with\nlanguage=\"ur\" hint.","operationId":"transcribe_voice_transcribe_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_transcribe_voice_transcribe_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/voice/speak":{"post":{"tags":["voice"],"summary":"Speak","description":"Convert text to speech using edge-tts (Microsoft neural voices, free).\nReturns an MP3 audio stream the browser can play directly.\n\nRoman Urdu text is handled well by ur-PK-AsadNeural.\nNumbers like \"Rs 50,000\" are read naturally.","operationId":"speak_voice_speak_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpeakRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/voice/chat":{"post":{"tags":["voice"],"summary":"Voice Chat","description":"Full pipeline: audio → Whisper → text (returned so UI can display it).\nThe caller then sends the text to /chat normally and calls /voice/speak\non the reply. This keeps the voice pipeline decoupled from the chat logic.","operationId":"voice_chat_voice_chat_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_voice_chat_voice_chat_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/whatsapp":{"post":{"tags":["whatsapp"],"summary":"Whatsapp Root","operationId":"whatsapp_root_whatsapp_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/whatsapp/webhook":{"post":{"tags":["whatsapp"],"summary":"Whatsapp Webhook","operationId":"whatsapp_webhook_whatsapp_webhook_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/wa/webhook":{"post":{"tags":["whatsapp-dispatcher"],"summary":"Whatsapp Dispatcher Webhook","description":"Unified WhatsApp webhook — receives all Twilio messages and\nroutes them to the correct Desibot based on active session.\nSet this as your Twilio sandbox webhook URL:\n    https://<ngrok>/wa","operationId":"whatsapp_dispatcher_webhook_wa_webhook_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/wa":{"post":{"tags":["whatsapp-dispatcher"],"summary":"Whatsapp Dispatcher Webhook","description":"Unified WhatsApp webhook — receives all Twilio messages and\nroutes them to the correct Desibot based on active session.\nSet this as your Twilio sandbox webhook URL:\n    https://<ngrok>/wa","operationId":"whatsapp_dispatcher_webhook_wa_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/chat":{"post":{"summary":"Chat","operationId":"chat_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/report/daily":{"get":{"summary":"Report Daily","operationId":"report_daily_report_daily_get","parameters":[{"name":"x-tenant-id","in":"header","required":false,"schema":{"type":"string","default":"default","title":"X-Tenant-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/report/weekly":{"get":{"summary":"Report Weekly","operationId":"report_weekly_report_weekly_get","parameters":[{"name":"x-tenant-id","in":"header","required":false,"schema":{"type":"string","default":"default","title":"X-Tenant-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/report/monthly":{"get":{"summary":"Report Monthly","operationId":"report_monthly_report_monthly_get","parameters":[{"name":"x-tenant-id","in":"header","required":false,"schema":{"type":"string","default":"default","title":"X-Tenant-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/report/schedule":{"get":{"summary":"Report Schedule","operationId":"report_schedule_report_schedule_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_transcribe_voice_transcribe_post":{"properties":{"audio":{"type":"string","format":"binary","title":"Audio"}},"type":"object","required":["audio"],"title":"Body_transcribe_voice_transcribe_post"},"Body_voice_chat_voice_chat_post":{"properties":{"audio":{"type":"string","format":"binary","title":"Audio"}},"type":"object","required":["audio"],"title":"Body_voice_chat_voice_chat_post"},"ChatIn":{"properties":{"message":{"type":"string","title":"Message"},"history":{"anyOf":[{"items":{"$ref":"#/components/schemas/HistoryMessage"},"type":"array"},{"type":"null"}],"title":"History","default":[]}},"type":"object","required":["message"],"title":"ChatIn"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HistoryMessage":{"properties":{"role":{"type":"string","title":"Role"},"content":{"type":"string","title":"Content"}},"type":"object","required":["role","content"],"title":"HistoryMessage"},"SpeakRequest":{"properties":{"text":{"type":"string","title":"Text"},"voice":{"type":"string","title":"Voice","default":"ur-PK-AsadNeural"},"rate":{"type":"string","title":"Rate","default":"+0%"},"volume":{"type":"string","title":"Volume","default":"+0%"}},"type":"object","required":["text"],"title":"SpeakRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}